Move shape_predictor_trainer to a separate header
Moved shape_predictor_trainer to its own header in order to support the
use of shape_predictor with DLIB_ISO_CPP_ONLY defined (i.e. without
threading API wrappers).
