+++ dlib/image_processing.h
@@ -20,6 +20,7 @@
 #include "image_processing/remove_unobtainable_rectangles.h"
 #include "image_processing/scan_fhog_pyramid.h"
 #include "image_processing/shape_predictor.h"
+#include "image_processing/shape_predictor_trainer.h"
 #include "image_processing/correlation_tracker.h"
 
 #endif // DLIB_IMAGE_PROCESSInG_H_h_
